+++ frontend/src/types/index.ts
@@ -41,6 +41,6 @@ export * from "./changeHistory";
 export * from "./semanticTypes";
 export * from "./iam";
 export * from "./misc";
-export * from "./custom-feature";
+export * from "./appProfile";
 
 export * from "./v1";
